+++ app/gui/src/dashboard/layouts/InfoMenu.tsx
@@ -1,11 +1,12 @@
 /** @file A menu containing info about the app. */
 import { LOGIN_PATH } from '#/appUtils'
-import { Popover, Text } from '#/components/AriaComponents'
+import { Popover } from '#/components/Dialog'
 import { Icon } from '#/components/Icon'
 import MenuEntry from '#/components/MenuEntry'
+import { Text } from '#/components/Text'
 import AboutModal from '#/modals/AboutModal'
 import { useAuth } from '#/providers/AuthProvider'
-import { useSetModal } from '#/providers/ModalProvider'
+import { setModal } from '#/providers/ModalProvider'
 import { useSessionAPI } from '#/providers/SessionProvider'
 import { useRouter, useText } from '$/providers/react'
 import { PRODUCT_NAME } from 'enso-common'
@@ -22,7 +23,6 @@ export default function InfoMenu(props: InfoMenuProps) {
   const { router } = useRouter()
   const { signOut } = useSessionAPI()
   const { session } = useAuth()
-  const { setModal } = useSetModal()
   const { getText } = useText()
 
   return (