+++ app/gui/src/project-view/components/FullscreenButton.vue
@@ -8,6 +8,7 @@ const fullscreen = defineModel<boolean>({ required: true })
 
 <template>
   <SvgButton
+    class="FullscreenButton"
     :name="fullscreen ? 'exit_fullscreen' : 'fullscreen'"
     :title="fullscreen ? 'Exit Fullscreen' : 'Fullscreen'"
     @activate="fullscreen = !fullscreen"