+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [Add option to browse cloud for folders][13117]
 - [File Browser Widget: Add ability to filter files by extension][13048]
 - [Add keyboard shortcuts for formatting documentation][13134]
+- [Allow selecting expected types for arguments of grouped components.][13161]
 
 [12774]: https://github.com/enso-org/enso/pull/12774
 [12778]: https://github.com/enso-org/enso/pull/12778
@@ -29,6 +30,7 @@
 [13117]: https://github.com/enso-org/enso/pull/13117
 [13048]: https://github.com/enso-org/enso/pull/13048
 [13134]: https://github.com/enso-org/enso/pull/13134
+[13161]: https://github.com/enso-org/enso/pull/13161
 
 #### Enso Standard Library